[mapper.default.fv] move out of playground, refs #15
